calculate days ago client side
Now that the entries in the feed are getting cached this data becomes
stale after a while, so in the cached fragment we'll display the actual
date (which will always be correct) and then update that to days ago
on the client. This is the approach that github also takes.

This also fixes #116