Add documentation for getting set up to use Docker
Added a section to the README documenting how to get started using
Docker and set up a `bin/docker_setup` script to make this easier.
I've made modifications to the /config/database.yml.example so that it can simply be
copied over and it will work for either development set up.